requirements.txt
requests==2.25.1
Flask==1.1.2
gunicorn==20.1.0
opencv-python==4.5.1.48
face-recognition==1.3.0
numpy==1.20.2
google-cloud-vision==2.3.0